requirements.txt
flask
flask-cors
spotipy
pandas
python-dotenv